Make Queries 2.0 compliant
